requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.6
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bandit==1.7.4
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.3
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
keyring==23.6.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.961
mypy-extensions==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.9.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2022.6.7
pkginfo==1.8.3
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylama==8.3.8
pylint==2.14.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytoolconfig==1.1.2
PyYAML==6.0
pyzmq==23.2.0
readme-renderer==35.0
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
rope==1.2.0
scikit-learn==1.1.1
scipy==1.8.1
SecretStorage==3.3.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
stevedore==3.5.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
twine==4.0.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
